Fix test_solver platform compatibility (again)
The tolerance has to be quite high because the max-to-min ratio of the
test data is large. This makes the comparison of the minimal values very
difficult due to platform dependent rounding. A baseline make on one
platform with not fit another one very well.

Lesson learned. Future baseline data will need to have a moderate ratio
of max-to-min values. The existing baselines should also be revised in
the same manner, but a more systematic approach to producing baseline
data needs to be devised first. For now, the high test tolerances will
serve, painful as they may be.