Fix complex hoppings with translational symmetry
This commit fixes a bug that was showcased by the tests added in the
previous commit.

The issue was that, before this commit, translational symmetry was only
being considered in one direction. It was assumed that the opposite
direction was taken care of by adding the both conjugates of the
hoppings, however the values were added without actually doing a complex
conjugate.

The fix in this commit reverses the logic: translational symmetry is
considered in all unique directions, even opposite ones, but the only
half the hoppings are added (non-conjugate). This works out naturally
as the Hermicity of the Hamiltonian matrix will take care of the
conjugates later.